Add data.sql file

+++ src/main/java/com/softserve/skillscope/talentInfo/TalentInfoRepository.java
@@ -3,5 +3,5 @@ package com.softserve.skillscope.talentInfo;
 import com.softserve.skillscope.talentInfo.model.entity.TalentInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TalentInfoRepository extends JpaRepository<TalentInfo, Integer> {
+public interface TalentInfoRepository extends JpaRepository<TalentInfo, Long> {
 }
